Project_Backtest_MultiStrategy/shared/setup_system/requirements.txt
kiteconnect==5.0.1
streamlit
pandas
mysql-connector-python
python-dotenv
openpyxl
